requirements.txt
openai>=1.0.0
tweepy>=4.14.0
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytz>=2024.1
tenacity>=8.0.0
structlog>=24.1.0
tiktoken>=0.5.2
better-profanity>=0.7.0
textblob>=0.17.1
nltk>=3.8.1
sentence-transformers>=2.5.0
numpy>=1.24.0
redis>=5.0.1
prometheus-client>=0.19.0
aioredis>=2.0.1
prometheus-async>=22.2.0
click>=8.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
fastapi-limiter>=0.1.5
python-jose[cryptography]>=3.3.0
httpx>=0.26.0 